+++ packages/yak-swc/yak_swc/tests/fixture/comments/skip
@@ -1,0 +1,1 @@
+Right now our Babel Plugin can't skip comments, but the SWC Plugin can.
