+++ packages/yak-swc/yak_swc/tests/fixture/cross-file-member/skip
@@ -1,0 +1,2 @@
+Right now the babel plugin rejects to processes the cross file `fonts.sm` constant
+The SWC plugin injects a reference to foreign file
