+++ packages/yak-swc/yak_swc/tests/fixture/cross-file-mixin/skip
@@ -1,0 +1,11 @@
+Right now the plugins are not aligned
+
+Babel properly handles this case:
+
+```
+export const Button = styled.button`
+  ${fonts.h1}
+`;
+```
+
+The SWC plugin fails as it can't determine wether fonts.h1 is a selector, a mixin or a variable.
