+++ packages/yak-swc/yak_swc/tests/fixture/css-mixin/skip
@@ -1,0 +1,1 @@
+Right now the Babel Plugin doesn't include the condition in the css variable name
