+++ packages/yak-swc/yak_swc/tests/fixture/no-yak-usage/skip
@@ -1,0 +1,2 @@
+The babel plugin does add an import right now
+This has to be fixed in the babel plugin
